Add TCR
Add a development script, tcr.sh, to try out a variation of
Test && Commit || Revert (TCR).

I'm trying a combination of TDD and TCR; I don't revert tests.
The tests are kept in a "dummy" ("red") commit which is automatically squashed with the subsequent "green" commit.
I'm considering changing this to `git stash` or similar instead of commiting the test, but it seems to work nicely in practice for now.

Thoughts so far:
* Miss having commit messages some times
* Irritating when a test/feature demands changing several lines of code:
** Reverting them all is annoying (kind of the point, I know)
** The lines could be changed one at a time _before_ the test, but "test first"?
* Very nice to have continuous feedback and automatic save/progress (commit)
* I'm doing much smaller commits than before. Like editor history. Fun.
* It seems to encourage fine grained tests.
* Encourages a very high level of test coverage.